requirements.txt
Flask==3.0.3
yt-dlp==2024.07.25
requests==2.32.3
gunicorn==22.0.0
psycopg2-binary
